requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-decouple==3.8
google-generativeai==0.4.1
openai==1.12.0
groq==0.4.2
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
django-filter==23.5
